passes: Refactor collect_derivatives
