+++ devito/passes/equations/__init__.py
@@ -1,1 +1,1 @@
-from .lowering import *  # noqa
+from .linearity import *  # noqa
